backend/protocol_rpc/requirements.txt
Flask-JSONRPC==3.0.1
Flask[async]==3.0.2
flask-socketio==5.3.6
Flask-Cors==4.0.0
psycopg2-binary>=2.9.9
requests==2.31.0
python-dotenv==1.0.1
asgiref==3.7.2
pydantic==2.6.4
numpy==1.26.4
pytest==8.1.1
colorama==0.4.6
debugpy==1.8.1
aiohttp==3.9.3
openai==1.16.1
SQLAlchemy[asyncio]==2.0.31
alembic==1.13.2
eth-account==0.13.1
eth-utils==4.1.1
sentence-transformers==3.0.1
Flask-SQLAlchemy==3.1.1
prometheus-client==0.20.0
